Make URIInput more generic
Don't assume there will be a form somewhere with a submit button.
Handle this by adding a new 'onSelect' prop, which is fired when
enter is pressed (in the default implementation). Generally, if
the underlying implementation is changed, the prop should be called
when the input is confirmed.

Selecting an element off the list with enter or a mouse click don't
trigger the prop. This is in purpose, to allow the user to select how
he wants to call the user.