Migrate main Blink component to ES6
This patch contains by far the uglyest JS - React code I've ever written.

Migrating this component to ES6 is not as simple as it should be because the
react-mini-router mixin contains React lifecycle methods, so we need to use
the onClass way for overriding mechanism. Also, since instance methods are
used as event listener we need to autobind them, but they are on the mixin,
so I created another mixin, to autobind the given methods. Ugly? Yes. Should
I lose my JS license? Surely. Does it work? Hell yes!

The 'routes' instance variable is now a class property, so I added the required
Babel transform for it to work.

Another learned lesson is that while generally componentWillMount can be
embedded in the constructor, no so in this case, because the mxin extends the
state, so this won't be visible inside the constructor.