Fixed passing notification center across components
There is a chicken and egg problem with passing around the notification center.  Since it's
a component and we reference it using 'refs', it needs to exist at least in the virtual DOM.
That is not the case for the first render, for example.  The React documentation also says
that `this.refs` should not be accessed from within render(), which means we shouldn't be
passing a ref as a prop.

So it looks like we have been dancing in undefined behavior territory for a while, and things
got worse after reunifying all notifications in the NotificationCenter.

Worry not!  This patch is The Solution.  Instead of passing the ref around we pass a getter function,
which returns the ref stored in the mail app component.  The main app component populates its
internal reference in componentDidMount, where the ref is guaranteed to exist, so it's all good now.