Add share button to hamburger menu
When doing so, I noticed the cplipboard feature stopped worling.  My suspicion is
the button is not in the DOM so a selector for it cannot be created.  After trying
a number of approaches I chose to ditch clopboard.js and do the basics ourselves.
It's just an ugly function in utils...