Reworked call header
Long URIs will break on the space and can be shortened on the next line with
ellipsis.

To handle cases the buttons are on top, headers been moved to a containing DIV
which has an absolute position. That way, when the buttons should be on top,
they are in the containing DIV and do not overlap the header. Position for the
buttonbar is now not needed anymore.

Fixes T225
Differential Revision: https://docs-new.sipthor.net/D49