Added conference participants video switching also on inactive stream event
On mobile it takes some time to end the stream, therefore stream.active could be still
true when the mayBe switch function is hit.
To fix this, the stream gets an eventListener for inactive. When this event happens
the mayBe function is called again. In this case active should be false.

Ref T261