Switch to audio only mode and disable camera if remote video track is recvonly
When answering a video call with webrtc and you have no camera, the video
stream is set to recvonly in the SDP. The remote sees it as a muted video
track. However since tracks can start muted, we need to check the a= from the
SDP. Since we send video only in one direction and the platform expects
symmetrical RTP streams, we make the stream 'symmetrical' by closing our camera
and switch the view to audio only mode.