Changed media to audio only for Safari on video conferences
It seems that in some versions of Safari the order of m lines in de SDP can
differ.(Tested on 12.1.2) If audio is first and the remote denies video, it
sort of works. If video is first and the remote denies video, you'll get an
error. If we build detection if video is working like in a normal call, it is
no guarantee that you don't hit the order issue.