Prevent loop in URI input when blurred from modal
If the input is blurred by opening, the state updated, that did trigger
didUpdate, that put focus, then it blurred again .....

Also if the value does not change, we don't need to set focus. If you do it
always on didUpdated means you can't ever lose focus which is wrong.